revert code enter logic & complete list logic

+++ package/extensions/d-block/dblock.ts
@@ -84,19 +84,20 @@ export const DBlock = Node.create<DBlockOptions>({
         } = editor.state;
 
         const parent = $head.node($head.depth - 1);
-
+        const currentNode = $head.node($head.depth);
         const headString = $head.toString();
         const nodePaths = headString.split('/');
+        const atTheStartOfText = from + 4;
 
         // Check if inside table
         const isInsideTable = nodePaths.some(path => path.includes('table'));
 
-        if (parent?.type.name !== 'dBlock') {
-          const isTaskList = nodePaths.some(path => path.includes('taskList'));
-          const isListOrTaskItem =
-            parent?.type.name === 'listItem' ||
-            parent?.type.name === 'taskItem';
+        const isListOrTaskItem =
+          parent?.type.name === 'listItem' || parent?.type.name === 'taskItem';
 
+        const isTaskList = parent?.type.name === 'taskItem';
+
+        if (parent?.type.name !== 'dBlock') {
           const isItemSelected = from !== to && isListOrTaskItem;
           // If a list item or task item is selected, delete it
           if (isItemSelected) {
@@ -109,24 +110,63 @@ export const DBlock = Node.create<DBlockOptions>({
           }
 
           const isCurrentItemEmpty =
-            parent?.lastChild?.textContent === '' &&
-            parent?.lastChild?.type.name === 'paragraph' &&
+            currentNode.textContent === '' &&
+            currentNode.type.name === 'paragraph' &&
             isListOrTaskItem;
 
-          if (isCurrentItemEmpty) {
-            return editor
-              .chain()
-              .deleteNode(isTaskList ? 'taskItem' : 'listItem')
-              .insertContentAt(from, {
-                type: 'dBlock',
-                content: [
-                  {
-                    type: 'paragraph',
-                  },
-                ],
-              })
-              .focus()
-              .run();
+          const grandParent = $head.node($head.depth - 2);
+
+          const isLastItemOfList =
+            $head.index($head.depth - 2) === grandParent.childCount - 1;
+          const isInList = ['bulletList', 'orderedList', 'taskList'].includes(
+            $head.node($head.depth - 2).type.name,
+          );
+
+          const isLastItemOfListWithoutContent = isLastItemOfList && isInList;
+          const isMiddleItemOfListWithoutContent =
+            !isLastItemOfList && isInList;
+          const isMultipleListItems =
+            $head.node($head.depth - 2).childCount > 1;
+
+          if (isCurrentItemEmpty && isMultipleListItems) {
+            if (isLastItemOfListWithoutContent) {
+              // Handle the case when it's the last empty item in the list
+              return editor
+                .chain()
+                .deleteNode(isTaskList ? 'taskItem' : 'listItem')
+                .insertContentAt(from, {
+                  type: 'dBlock',
+                  content: [
+                    {
+                      type: 'paragraph',
+                    },
+                  ],
+                })
+                .focus()
+                .run();
+            } else if (isMiddleItemOfListWithoutContent) {
+              // Handle the case when it's an empty middle item in the list
+              return editor
+                .chain()
+                .deleteCurrentNode()
+                .insertContentAt(from, {
+                  type: isTaskList ? 'taskItem' : 'listItem',
+                  content: [
+                    {
+                      type: 'paragraph',
+                    },
+                  ],
+                })
+                .focus(atTheStartOfText)
+                .run();
+            } else {
+              // Handle the case for the first empty item in the list
+              return editor
+                .chain()
+                .liftListItem(isTaskList ? 'taskItem' : 'listItem')
+                .focus()
+                .run();
+            }
           } else {
             return false;
           }
@@ -152,7 +192,9 @@ export const DBlock = Node.create<DBlockOptions>({
         const content = doc.slice(from, currentActiveNodeTo)?.toJSON().content;
 
         try {
-          const atTheStartOfText = from + 4;
+          if (currentActiveNodeType === 'codeBlock') {
+            return editor.chain().newlineInCode().focus().run();
+          }
 
           if (['columns', 'heading'].includes(currentActiveNodeType)) {
             return editor
